requirements.txt
torch
accelerate
monai
timm
easydict
tqdm
torchvision
pycocotools
pillow